* nuspec template improvements * AssemblyInfo.cs template fixes to allow setting the assembly guid when building for other than NETSTANDARD platforms * build script imptovements

+++ core/src/main/build.sh
@@ -1,12 +1,30 @@
-if [ -z "./packages" ]; then
-  ./paket.sh update
+msbuild="../../../submodules/btw/msbuild.sh"
+paket='.paket/paket.sh'
+project='Axle.Core'
+
+$paket update
+if [ $? -ne 0 ]; then
+  read -rsp "Press [Enter] to quit"
+  echo ""
+  exit
 fi
 
-msbuild="../../../submodules/btw/msbuild.sh"
-
-# clean-up obj directory to remove .net standard/core files (project.json) which cause issues with legacy framework version builds
 rm -rf obj/
-dotnet restore Axle.Core.csproj
-$msbuild Axle.Core.csproj /tv:15.0
-# pack
-#$msbuild Axle.Core.dist.csproj
+dotnet restore $project.csproj
+if [ $? -ne 0 ]; then
+  read -rsp "Press [Enter] to quit"
+  echo ""
+  exit
+fi
+$msbuild $project.csproj
+if [ $? -ne 0 ]; then
+  read -rsp "Press [Enter] to quit"
+  echo ""
+  exit
+fi
+$msbuild $project.dist.csproj
+if [ $? -ne 0 ]; then
+  read -rsp "Press [Enter] to quit"
+  echo ""
+  exit
+fi
